requirements.txt
torchfold
psqlparse==1.0rc7
psycopg2-binary
torch
git+https://github.com/mhoangvslev/torchfold.git@master
SPARQLWrapper==1.8.5